refactor: one phase correction pivot point
